chore: move errors in their own module
