+++ crates/assets/src/lib.rs
@@ -28,7 +28,10 @@ pub mod asset;
 pub mod chain;
 pub mod v1;
 
+mod errors;
+
 // Re-export relevant types
 pub use asset::*;
 pub use chain::*;
+pub use errors::*;
 pub use v1::*;
